SMBIOS and kernel: clock rate converted to u64

+++ include/device_info/SMBIOS.cpp
@@ -14,7 +14,7 @@
 
 smbios_t* smbios = nullptr;
 smbios_processor_info_t* cpu_info = nullptr;
-i32 clock_rate = 0;
+u64 clock_rate = 0;
 
 
 smbios_t search_for_SMBIOS()
@@ -99,7 +99,7 @@ u64 SMBIOS_get_CPU_clock_rate_hz()
     {
         SMBIOS_populate_cpu_info();
     }
-    clock_rate = static_cast<i32>(cpu_info->current_speed) * 1000000;
+    clock_rate = static_cast<u64>(cpu_info->current_speed) * 1000000;
     return clock_rate;
 }
